AzureVirtualMachineFilter-GUI_Based_Tool
What does this tool do ?

This is a tiny tool which provides a simple UI interface for performing
– view, filter and export operations on an Azure VM repository based on
any parameter including Subnet, IPAddress, PowerState, AvailabilitySet,
Size, Tags etc which are currently not available through the Azure
Portal or through a simple PowerShell command.

Why this tool ?

Let’s say you want to get a list of all VMs in your subscription from a
particular “subnet” or a particular “availability set”

• as of today there is no option to view that information using the
Azure Portal ( and )
• there is no easy way to fetch that information even using PowerShell –
you have to run a merge array of sorts with the output of two commands.

This tool aims to provide an easy to use, intuitive UI based tool to
help bridge this gap