NSG Tool ( Azure Network Security Group UI Tool ) - EXE
(moved from nsgtool.codeplex.com)

EXE Version 

This is the "ridiculously" simple UI tool for managing Network Security Groups (NSGs) in Azure. You can manage NSGs and NSG rules across subscription including both the Classic (ASM) model of Azure and ARM using a single intuitive UI pane. You will also be able to view the associated NSGs for the subnets and VMs, export to a CSV and the best of all - filter the rules at the click of a button

Description - https://youtu.be/vgnSLuf_zj0